Doh, again.

+++ tools/generate_indexes.py
@@ -52,7 +52,9 @@ for root, dirs, files in os.walk('.'):
             dirs=diritems
         )
         outpath = os.path.join(root, 'index.html')
-        if (not os.path.exists(outpath)):
+        if (not os.path.exists(outpath) and os.access(outpath, os.W_OK)):
             with open(outpath, 'w', encoding='utf-8') as out:
                 out.write(output)
             print(f"✅ {outpath} erstellt.")
+        else:
+            print(f"{outpath} nicht erstellt. Rechte vorhanden: {os.access(outpath, os.W_OK)}")
